requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-pptx
playwright
beautifulsoup4
python-dotenv
asyncio
cloudinary